Move helpers under the Win32::TaskScheduler namespace
The `Windows` namespace was unnecessary and there were helpers outside
of the namespace of this library that caused conflicts with other
libraries, i.e. `SID`.
